requirements.txt
boto3==1.35.88
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1